requirements.txt
Flask
faiss-cpu
numpy
ollama
pdfplumber
python-docx
openpyxl
werkzeug
